refactor(arch): Phase 5 — PR #486 리뷰 7차 수정
- B1: deleteAccountAction normalizeEmail 적용
- B2: OAuth route handlers deep import → barrel (callback + start)
- S2: ESLint features cross-import 주석에 구체적 슬라이스 쌍 명시
