+++ eslint.config.mjs
@@ -99,10 +99,9 @@ const eslintConfig = defineConfig([
                             ],
                         },
                         {
-                            // Phase 5: features 간 cross-import 허용. auth-signup이 auth-email-verification을,
-                            // auth-oauth-consent가 auth-signup을, account-delete가 auth-oauth를 참조하는 등
-                            // auth 슬라이스 간 의존이 불가피.
-                            // TODO(Phase 7): widgets 레이어 분리 시 공유 로직을 widgets로 올려 features 간 직접 의존 제거.
+                            // Phase 5 임시 허용 — auth 슬라이스 간 cross-import.
+                            // 허용 쌍: auth-signup → auth-email-verification, auth-oauth-consent → auth-signup
+                            // TODO(Phase 7): 공유 로직을 entities/shared로 추출하여 해소.
                             from: 'features',
                             allow: [
                                 'features',
